refactor(components): Move chat components to features directory
Consolidate `Message`, `MessageList`, and `ChatInput` under `aiChat` feature
directory and remove from common exports to organize by feature rather
than component type.
