--- frontend/src/components/common/index.js
@@ -1,3 +1,0 @@
-export { Message } from './Message'
-export { MessageList } from './MessageList'
-export { ChatInput } from './ChatInput'
