+++ frontend/src/components/features/aiChat/ChatInput.jsx
