+++ frontend/src/components/features/aiChat/index.js
@@ -1,0 +1,4 @@
+export * from './aiChat'
+export * from './Message'
+export * from './MessageList'
+export * from './ChatInput'
